+++ Dominio/Carrito.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 
 namespace Dominio
 {
-    public class Carrito : EntidadBase<long>
+    public class Carrito : EntidadBase<int>
     {
         public virtual List<Articulo> Articulos { get; set; }
         public virtual decimal Monto { get; set; }
